bump appveyor image to Visual Studio 2022
hopefully resolves crazy audio stuttering for appveyor builds.
this seems to mainly happen on appveyor builds and it appears to be fine on azure builds. the only difference i could spot is that appveyor was using `Visual Studio 15 2017` while azure is using a slightly more up to date version `Visual Studio 16 2019`.